Refactor module layout and exports
